# api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LongLink Platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/storages": {
      "post": {
        "summary": "Create Storage Registry",
        "description": "Register one Exoscale SOS backend.",
        "operationId": "create_storage_registry_api_v1_storages_post",
        "parameters": [
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageRegistryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageRegistryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Storage Registries",
        "description": "Return all registered storage backends.",
        "operationId": "list_storage_registries_api_v1_storages_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StorageRegistryResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storages/{registry_id}": {
      "get": {
        "summary": "Get Storage Registry",
        "description": "Return one storage backend registration.",
        "operationId": "get_storage_registry_api_v1_storages__registry_id__get",
        "parameters": [
          {
            "name": "registry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Registry Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageRegistryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Storage Registry",
        "description": "Delete one unused Exoscale SOS backend registration.",
        "operationId": "delete_storage_registry_api_v1_storages__registry_id__delete",
        "parameters": [
          {
            "name": "registry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Registry Id"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "summary": "Get Me",
        "description": "Return the authenticated user's details.",
        "operationId": "get_me_api_v1_me_get",
        "parameters": [
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Me",
        "description": "Update the authenticated user's details.",
        "operationId": "patch_me_api_v1_me_patch",
        "parameters": [
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/organizations": {
      "get": {
        "summary": "Get My Organizations",
        "description": "Return the authenticated user's organization memberships.",
        "operationId": "get_my_organizations_api_v1_me_organizations_get",
        "parameters": [
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserOrganizationMembership"
                  },
                  "title": "Response Get My Organizations Api V1 Me Organizations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "summary": "List Users",
        "description": "Return all user summaries for administrator views.",
        "operationId": "list_users_api_v1_users_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "longlink_auth",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longlink Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UserSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrganizationIdentity": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "avatar": {
            "type": "string",
            "title": "Avatar"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "avatar"
        ],
        "title": "OrganizationIdentity",
        "description": "Represent a compact Organization in nested API responses."
      },
...
      "OrganizationRoles": {
        "type": "string",
        "enum": [
          "read",
          "write",
          "maintain",
          "admin",
          "owner"
        ],
        "title": "OrganizationRoles",
        "description": "Supported organization membership roles."
      },
...
      "Page_StorageRegistryResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StorageRegistryResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[StorageRegistryResponse]"
      },
      "Page_UserSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[UserSummary]"
      },
...
      "StorageRegistryCreate": {
        "properties": {
          "endpoint_url": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Endpoint Url"
          },
          "access_key_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Access Key Id"
          },
          "secret_access_key": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Secret Access Key"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "endpoint_url",
          "access_key_id",
          "secret_access_key",
          "name"
        ],
        "title": "StorageRegistryCreate",
        "description": "Validate one storage registry creation payload."
      },
      "StorageRegistryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "endpoint_url": {
            "type": "string",
            "title": "Endpoint Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "endpoint_url"
        ],
        "title": "StorageRegistryResponse",
        "description": "Describe one Exoscale SOS backend without exposing Platform credentials."
      },
...
      "UserOrganizationMembership": {
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationIdentity"
          },
          "role": {
            "$ref": "#/components/schemas/OrganizationRoles"
          }
        },
        "type": "object",
        "required": [
          "organization",
          "role"
        ],
        "title": "UserOrganizationMembership",
        "description": "Represent one current-user Organization membership."
      },
      "UserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "avatar": {
            "type": "string",
            "title": "Avatar"
          },
          "administrator": {
            "type": "boolean",
            "title": "Administrator"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "avatar",
          "administrator"
        ],
        "title": "UserSummary",
        "description": "Represent a compact user object in nested responses."
      },
      "UserUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "Payload to update mutable user profile fields."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}